+++ crates/jlrs/tests/array_tests_impl/array_bits_data_mut.rs
@@ -6,7 +6,7 @@ pub(crate) mod tests {
                 TypedRankedArray,
                 data::accessor::{Accessor, AccessorMut},
             },
-            types::construct_type::ConstantBool,
+            types::construct_type::constants::ConstantBool,
         },
         prelude::*,
     };