+++ crates/jlrs/tests/array_tests_impl/array_union_data_mut.rs
@@ -3,7 +3,7 @@ pub(crate) mod tests {
     use jlrs::{
         data::{
             managed::array::data::accessor::AccessorMut,
-            types::construct_type::UnionTypeConstructor,
+            types::construct_type::union_types::UnionTypeConstructor,
         },
         prelude::*,
     };