+++ tests/julia_module_test/src/array.rs
@@ -6,7 +6,7 @@ use jlrs::{
         },
         types::{
             abstract_type::Integer,
-            construct_type::{ArrayTypeConstructor, ConstantIsize},
+            construct_type::{array::ArrayTypeConstructor, constants::ConstantIsize},
         },
     },
     prelude::{
@@ -58,50 +58,64 @@ pub fn takes_typed_ranked_array(a: TypedRankedArray<u32, 1>) -> usize {
 
 // Array return type
 pub unsafe fn returns_array(dt: DataType) -> ArrayRet {
-    let weak_handle = weak_handle_unchecked!();
-    Array::new_for(weak_handle, dt.as_value(), [2, 2])
-        .unwrap()
-        .leak()
+    unsafe {
+        let weak_handle = weak_handle_unchecked!();
+        Array::new_for(weak_handle, dt.as_value(), [2, 2])
+            .unwrap()
+            .leak()
+    }
 }
 
 pub unsafe fn returns_rank0_array(dt: DataType) -> RankedArrayRet<0> {
-    let weak_handle = weak_handle_unchecked!();
-    RankedArray::<0>::new_for(weak_handle, dt.as_value(), [])
-        .unwrap()
-        .leak()
+    unsafe {
+        let weak_handle = weak_handle_unchecked!();
+        RankedArray::<0>::new_for(weak_handle, dt.as_value(), [])
+            .unwrap()
+            .leak()
+    }
 }
 
 pub unsafe fn returns_rank1_array(dt: DataType) -> RankedArrayRet<1> {
-    let weak_handle = weak_handle_unchecked!();
-    RankedArray::<1>::new_for(weak_handle, dt.as_value(), [2])
-        .unwrap()
-        .leak()
+    unsafe {
+        let weak_handle = weak_handle_unchecked!();
+        RankedArray::<1>::new_for(weak_handle, dt.as_value(), [2])
+            .unwrap()
+            .leak()
+    }
 }
 
 pub unsafe fn returns_rank2_array(dt: DataType) -> RankedArrayRet<2> {
-    let weak_handle = weak_handle_unchecked!();
-    RankedArray::<2>::new_for(weak_handle, dt.as_value(), [2, 2])
-        .unwrap()
-        .leak()
+    unsafe {
+        let weak_handle = weak_handle_unchecked!();
+        RankedArray::<2>::new_for(weak_handle, dt.as_value(), [2, 2])
+            .unwrap()
+            .leak()
+    }
 }
 
 pub unsafe fn returns_rank3_array(dt: DataType) -> RankedArrayRet<3> {
-    let weak_handle = weak_handle_unchecked!();
-    RankedArray::<3>::new_for(weak_handle, dt.as_value(), [2, 2, 2])
-        .unwrap()
-        .leak()
+    unsafe {
+        let weak_handle = weak_handle_unchecked!();
+        RankedArray::<3>::new_for(weak_handle, dt.as_value(), [2, 2, 2])
+            .unwrap()
+            .leak()
+    }
 }
 
 pub unsafe fn returns_typed_array() -> TypedArrayRet<f32> {
-    let weak_handle = weak_handle_unchecked!();
-    TypedArray::<f32>::new(weak_handle, [2, 2]).unwrap().leak()
+    unsafe {
+        let weak_handle = weak_handle_unchecked!();
+        TypedArray::<f32>::new(weak_handle, [2, 2]).unwrap().leak()
+    }
 }
 
 pub unsafe fn returns_typed_rank2_array() -> TypedRankedArrayRet<f32, 2> {
-    let weak_handle = weak_handle_unchecked!();
-    TypedRankedArray::<f32, 2>::new(weak_handle, [2, 2])
-        .unwrap()
-        .leak()
+    unsafe {
+        let weak_handle = weak_handle_unchecked!();
+        TypedRankedArray::<f32, 2>::new(weak_handle, [2, 2])
+            .unwrap()
+            .leak()
+    }
 }
 
 // Generic arrays